Refactored GameControl.cs
I was finding it difficult to understand what everything was doing and how to fit in the question pop-up code, so while analyzing code to understand it I found a lot of places where repeated code could be generalized and remove a lot of un-needed calls.
I have tested everything after making the changes and all functions appear to be working as before. I have left all the original code in comments for the time being for review but will remove them in a later commit